+++ killstats/tests/test_api/_killstasts_api.py
@@ -1,6 +1,23 @@
 KillstatsMonth = [
     {
-        "kills": [],
+        "kills": [
+            {
+                "title": None,
+                "killmail_id": 119303075,
+                "name": "Garthd",
+                "character_id": 1004,
+                "character_name": "Garthd",
+                "portrait": "https://images.evetech.net/types/29340/icon?size=64",
+                "corporation_id": 2001,
+                "alliance_id": 3001,
+                "ship": 29340,
+                "ship_name": "Osprey Navy Issue",
+                "hash": "1185d04b1b6c6667167aa7d65245a49d19d44275",
+                "totalValue": 1000,
+                "date": "2024-07-07T21:38:58Z",
+                "count": 0,
+            }
+        ],
         "losses": [
             {
                 "title": "None",
@@ -19,7 +36,7 @@ KillstatsMonth = [
                 "count": 0,
             }
         ],
-        "totalKills": 0,
+        "totalKills": 1000,
         "totalLoss": 76734851,
         "shame": [
             {
@@ -83,6 +100,49 @@ KillstatsMonth = [
                 "portrait": "https://images.evetech.net/characters/1001/portrait?size=256",
                 "count": 1,
                 "loss": True,
+            },
+            {
+                "title": "Top Kill:",
+                "killmail_id": 119303075,
+                "name": "Garthd",
+                "character_id": 1004,
+                "character_name": "Garthd",
+                "portrait": "https://images.evetech.net/types/29340/icon?size=64",
+                "corporation_id": 2001,
+                "alliance_id": 3001,
+                "ship": 29340,
+                "ship_name": "Osprey Navy Issue",
+                "hash": "1185d04b1b6c6667167aa7d65245a49d19d44275",
+                "totalValue": 1000,
+                "date": "2024-07-07T21:38:58Z",
+                "count": 0,
+            },
+            {
+                "title": "Top Ship:",
+                "type": "count",
+                "ship": 17634,
+                "ship_name": "Caracal Navy Issue",
+                "portrait": "https://images.evetech.net/types/17634/icon?size=256",
+                "count": 1,
+                "loss": False,
+            },
+            {
+                "title": "Top Killer: Gneuten",
+                "type": "count",
+                "character_id": 1001,
+                "character_name": "Gneuten",
+                "portrait": "https://images.evetech.net/characters/1001/portrait?size=256",
+                "count": 1,
+                "loss": False,
+            },
+            {
+                "title": "Alltime Killer: Gneuten",
+                "type": "count",
+                "character_id": 1001,
+                "character_name": "Gneuten",
+                "portrait": "https://images.evetech.net/characters/1001/portrait?size=256",
+                "count": 1,
+                "loss": False,
             },
         ],
     }