+++ killstats/tests/test_api/test_killstatsapi.py
@@ -7,7 +7,7 @@ from django.test import TestCase
 from app_utils.testing import create_user_from_evecharacter
 
 from killstats.api.killboard import KillboardApiEndpoints
-from killstats.tests.test_api.killstasts_api import KillstatsMonth
+from killstats.tests.test_api._killstasts_api import KillstatsMonth
 from killstats.tests.testdata.load_allianceauth import load_allianceauth
 from killstats.tests.testdata.load_killstats import load_killstats_all
 
@@ -34,11 +34,13 @@ class ManageApiJournalCharEndpointsTest(TestCase):
         cls.manage_api_endpoints = KillboardApiEndpoints(api=cls.api)
 
     def test_killstats_killboard_api(self):
+        self.maxDiff = None
         # given
         self.client.force_login(self.user)
         url = "/killstats/api/killboard/month/7/year/2024/"
         # when
         response = self.client.get(url)
+        print(response.json())
         # then
         expected_data = KillstatsMonth
         self.assertEqual(response.status_code, 200)